+++ src/main/java/core/mvc/tobe/AnnotationHandlerMapping.java
@@ -1,5 +1,6 @@
 package core.mvc.tobe;
 
+import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import core.annotation.web.Controller;
 import core.annotation.web.RequestMethod;
@@ -7,12 +8,14 @@ import core.di.context.ApplicationContext;
 import core.mvc.HandlerMapping;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
 import javax.servlet.http.HttpServletRequest;
 import java.lang.annotation.Annotation;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -20,53 +23,23 @@ import java.util.stream.Collectors;
 public class AnnotationHandlerMapping implements HandlerMapping {
     private static final Logger logger = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
 
-    private ApplicationContext applicationContext;
-    private HandlerConverter handlerConverter;
-    private ExceptionHandlerConverter exceptionHandlerConverter;
+    private final ApplicationContext applicationContext;
+    private final List<HandlerConverter> handlerConverters = Lists.newArrayList();
 
-    private Map<HandlerKey, HandlerExecution> handlerExecutions = Maps.newHashMap();
+    private final Map<HandlerKey, HandlerExecution> handlerExecutions = Maps.newHashMap();
 
-    public AnnotationHandlerMapping(ApplicationContext applicationContext, HandlerConverter handlerConverter,
-                                    ExceptionHandlerConverter exceptionHandlerConverter) {
+    public AnnotationHandlerMapping(ApplicationContext applicationContext, HandlerConverter... handlerConverters) {
         this.applicationContext = applicationContext;
-        this.handlerConverter = handlerConverter;
-        this.exceptionHandlerConverter = exceptionHandlerConverter;
+
+        Collections.addAll(this.handlerConverters, handlerConverters);
     }
 
     public void initialize() {
-        Map<Class<?>, Object> controllers = getControllers(applicationContext);
-        handlerExecutions.putAll(handlerConverter.convert(controllers));
+        for (HandlerConverter handlerConverter : handlerConverters) {
+            handlerExecutions.putAll(handlerConverter.convert(applicationContext));
+        }
 
-        Map<Class<?>, Object> exceptionHandleResolvers = getExceptionHandlerResolver(applicationContext);
-        handlerExecutions.putAll(exceptionHandlerConverter.convert(exceptionHandleResolvers));
         logger.info("Initialized AnnotationHandlerMapping!");
-    }
-
-    private Map<Class<?>, Object> getControllers(ApplicationContext ac) {
-        Map<Class<?>, Object> controllers = Maps.newHashMap();
-        for (Class<?> clazz : ac.getBeanClasses()) {
-            Annotation annotation = clazz.getAnnotation(Controller.class);
-            if (annotation != null) {
-                controllers.put(clazz, ac.getBean(clazz));
-            }
-        }
-        return controllers;
-    }
-
-    private Map<Class<?>, Object> getExceptionHandlerResolver(ApplicationContext ac) {
-        return ac.getBeanClasses()
-                .stream()
-                .filter(this::supportedException)
-                .collect(Collectors.toMap(Function.identity(), ac::getBean));
-    }
-
-    private boolean supportedException(Class<?> clazz) {
-        if (clazz.isAnnotationPresent(ControllerAdvice.class)) {
-            return true;
-        }
-
-        return Arrays.stream(clazz.getDeclaredMethods())
-                .anyMatch(method -> method.isAnnotationPresent(ExceptionHandler.class));
     }
 
     @Override
@@ -79,7 +52,7 @@ public class AnnotationHandlerMapping implements HandlerMapping {
 
     @Override
     public Object getHandler(Throwable e) {
-        logger.debug("Throwable : {}", e);
+        logger.debug("Throwable : {}", e.getMessage());
         return getHandlerInternal(new ExceptionHandlerKey(new Class[]{e.getClass()}));
     }
