+++ src/main/java/core/mvc/tobe/ExceptionHandlerConverter.java
@@ -1,6 +1,7 @@
 package core.mvc.tobe;
 
 import com.google.common.collect.Lists;
+import core.di.context.ApplicationContext;
 import core.mvc.tobe.support.ArgumentResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,23 +15,30 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
-public class ExceptionHandlerConverter {
+public class ExceptionHandlerConverter implements HandlerConverter {
     private static final Logger logger = LoggerFactory.getLogger(ExceptionHandlerConverter.class);
 
     private List<ArgumentResolver> argumentResolvers = Lists.newArrayList();
 
     private static final ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
 
+    @Override
     public void setArgumentResolvers(List<ArgumentResolver> argumentResolvers) {
         this.argumentResolvers.addAll(argumentResolvers);
     }
 
+    @Override
     public void addArgumentResolver(ArgumentResolver argumentResolver) {
         this.argumentResolvers.add(argumentResolver);
     }
 
-    public Map<HandlerKey, HandlerExecution> convert(Map<Class<?>, Object> exceptionHandlers) {
+    @Override
+    public Map<HandlerKey, HandlerExecution> convert(ApplicationContext ac) {
+        Map<Class<?>, Object> exceptionHandlers = getExceptionHandlerResolver(ac);
+
         Map<HandlerKey, HandlerExecution> handlers = new HashMap<>();
         Set<Class<?>> exceptionHandlerClass = exceptionHandlers.keySet();
         for (Class<?> exceptionHandler : exceptionHandlerClass) {
@@ -39,6 +47,18 @@ public class ExceptionHandlerConverter {
         }
 
         return handlers;
+    }
+
+    private Map<Class<?>, Object> getExceptionHandlerResolver(ApplicationContext ac) {
+        return ac.getBeanClasses()
+                .stream()
+                .filter(this::supportedException)
+                .collect(Collectors.toMap(Function.identity(), ac::getBean));
+    }
+
+    private boolean supportedException(Class<?> clazz) {
+        return Arrays.stream(clazz.getDeclaredMethods())
+                .anyMatch(method -> method.isAnnotationPresent(ExceptionHandler.class));
     }
 
     private void addHandlerExecution(Map<HandlerKey, HandlerExecution> handlers, final Object target, Method[] methods) {
