+++ src/main/java/core/mvc/tobe/HandlerConverter.java
@@ -1,57 +1,16 @@
 package core.mvc.tobe;
 
-import com.google.common.collect.Lists;
-import core.annotation.web.RequestMapping;
+import core.di.context.ApplicationContext;
 import core.mvc.tobe.support.ArgumentResolver;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.core.LocalVariableTableParameterNameDiscoverer;
-import org.springframework.core.ParameterNameDiscoverer;
 
-import java.lang.reflect.Method;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
-public class HandlerConverter {
+public interface HandlerConverter {
 
-    private static final Logger logger = LoggerFactory.getLogger(HandlerConverter.class);
+    void setArgumentResolvers(List<ArgumentResolver> argumentResolvers);
 
-    private List<ArgumentResolver> argumentResolvers = Lists.newArrayList();
+    void addArgumentResolver(ArgumentResolver argumentResolver);
 
-    private static final ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
-
-    public void setArgumentResolvers(List<ArgumentResolver> argumentResolvers) {
-        this.argumentResolvers.addAll(argumentResolvers);
-    }
-
-    public void addArgumentResolver(ArgumentResolver argumentResolver) {
-        this.argumentResolvers.add(argumentResolver);
-    }
-
-    public Map<HandlerKey, HandlerExecution> convert(Map<Class<?>, Object> controllers) {
-        Map<HandlerKey, HandlerExecution> handlers = new HashMap<>();
-        Set<Class<?>> controllerClazz = controllers.keySet();
-        for (Class<?> controller : controllerClazz) {
-            Object target = controllers.get(controller);
-            addHandlerExecution(handlers, target, controller.getMethods());
-        }
-
-        return handlers;
-    }
-
-    private void addHandlerExecution(Map<HandlerKey, HandlerExecution> handlers, final Object target, Method[] methods) {
-        Arrays.stream(methods)
-                .filter(method -> method.isAnnotationPresent(RequestMapping.class))
-                .forEach(method -> {
-                    RequestMapping requestMapping = method.getAnnotation(RequestMapping.class);
-                    RequestHandlerKey requestHandlerKey = new RequestHandlerKey(requestMapping.value(), requestMapping.method());
-                    HandlerExecution handlerExecution = new HandlerExecution(nameDiscoverer, argumentResolvers, target, method);
-                    handlers.put(requestHandlerKey, handlerExecution);
-                    logger.info("Add - method: {}, path: {}, HandlerExecution: {}", requestMapping.method(), requestMapping.value(), method.getName());
-                });
-    }
-
+    Map<? extends HandlerKey,? extends HandlerExecution> convert(ApplicationContext applicationContext);
 }
