+++ src/main/java/next/config/MyConfiguration.java
@@ -5,7 +5,7 @@ import core.annotation.ComponentScan;
 import core.annotation.Configuration;
 import core.jdbc.JdbcTemplate;
 import core.mvc.tobe.ExceptionHandlerConverter;
-import core.mvc.tobe.HandlerConverter;
+import core.mvc.tobe.RequestHandlerConverter;
 import core.mvc.tobe.support.*;
 import next.security.LoginUserArgumentResolver;
 import org.apache.commons.dbcp2.BasicDataSource;
@@ -38,17 +38,16 @@ public class MyConfiguration {
     public ExceptionHandlerConverter exceptionHandlerConverter() {
         ExceptionHandlerConverter exceptionHandlerConverter = new ExceptionHandlerConverter();
         exceptionHandlerConverter.setArgumentResolvers(defaultArgumentResolvers());
-        exceptionHandlerConverter.addArgumentResolver(loginUserArgumentResolver());
 
         return exceptionHandlerConverter;
     }
 
     @Bean
-    public HandlerConverter handlerConverter() {
-        HandlerConverter handlerConverter = new HandlerConverter();
-        handlerConverter.setArgumentResolvers(defaultArgumentResolvers());
-        handlerConverter.addArgumentResolver(loginUserArgumentResolver());
-        return handlerConverter;
+    public RequestHandlerConverter handlerConverter() {
+        RequestHandlerConverter requestHandlerConverter = new RequestHandlerConverter();
+        requestHandlerConverter.setArgumentResolvers(defaultArgumentResolvers());
+        requestHandlerConverter.addArgumentResolver(loginUserArgumentResolver());
+        return requestHandlerConverter;
     }
 
     LoginUserArgumentResolver loginUserArgumentResolver() {
